Adding WearTouchPad extension
WearTouchPad is a GearVRf extension. It is an app that should be installed on the phone.
This app has a companion Android wear app. The mobile app gets all the touchpad events from
the companion Android wear app. This extension is used by the GVRAndroidWearTouchpad for getting
touchpad events from the Android wear watch and delivering it to the GVRActivity.
The GVRf/Extensions/WearTouchPad/WearTouchPad.apk apk file is a combined apk file that will install the app on
the phone as well as the connected Android wear watch. This is included for easily installing the app on the watch and the phone.